separate class demo and real demo
