+++ ote_sdk/ote_sdk/usecases/exportable_code/demo/setup.py
@@ -42,5 +42,4 @@ setup(
         packages[0]: ["*.json"],
     },
     install_requires=required,
-    entry_points={"console_scripts": [f"{packages[0]}={packages[0]}.sync:main"]},
 )
